+++ move.c
@@ -6,13 +6,13 @@
 /*   By: phudyka <phudyka@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 17:39:07 by phudyka           #+#    #+#             */
-/*   Updated: 2022/11/28 10:54:49 by phudyka          ###   ########.fr       */
+/*   Updated: 2022/12/07 10:30:03 by phudyka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-static void	rotation_a(t_stack **stack_a, long *costa)
+static void	rotation_a(t_stack **stack_a, int *costa)
 {
 	if ((*costa) > 0)
 	{
@@ -26,7 +26,7 @@ static void	rotation_a(t_stack **stack_a, long *costa)
 	}
 }
 
-static void	rotation_b(t_stack **stack_b, long *cosby)
+static void	rotation_b(t_stack **stack_b, int *cosby)
 {
 	if ((*cosby) > 0)
 	{
@@ -40,21 +40,21 @@ static void	rotation_b(t_stack **stack_b, long *cosby)
 	}
 }
 
-static void	double_rotation(t_stack **stack_a, t_stack **stack_b, long *costa, long *cosby)
+static void	double_rotation(t_stack **stack_a, t_stack **stack_b, int *costa, int *cosby)
 {
 	rr(stack_a, stack_b);
 	(*costa)--;
 	(*cosby)--;
 }
 
-static void	double_reverse_rotation(t_stack **stack_a, t_stack **stack_b, long *costa, long *cosby)
+static void	double_reverse_rotation(t_stack **stack_a, t_stack **stack_b, int *costa, int *cosby)
 {
 	rrr(stack_a, stack_b);
 	(*costa)++;
 	(*cosby)++;
 }
 
-void	master_rotation(t_stack **stack_a, t_stack  **stack_b, long costa, long cosby)
+void	master_rotation(t_stack **stack_a, t_stack  **stack_b, int costa, int cosby)
 {
 	while (costa != 0 || cosby != 0)
 	{
